// frontend/src/api/models/createExtracurricularRequestStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inspirate Consulting API
 * API for the Inspirate Consulting application
 * OpenAPI spec version: 1.0.0
 */

export type CreateExtracurricularRequestStatus = typeof CreateExtracurricularRequestStatus[keyof typeof CreateExtracurricularRequestStatus];


export const CreateExtracurricularRequestStatus = {
  doing: 'doing',
  have_done: 'have_done',
} as const;