// frontend/src/api/models/createExtracurricularRequestType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inspirate Consulting API
 * API for the Inspirate Consulting application
 * OpenAPI spec version: 1.0.0
 */

export type CreateExtracurricularRequestType = typeof CreateExtracurricularRequestType[keyof typeof CreateExtracurricularRequestType];


export const CreateExtracurricularRequestType = {
  maintenance: 'maintenance',
  investment: 'investment',
} as const;