// frontend/src/api/models/createExtracurricularRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inspirate Consulting API
 * API for the Inspirate Consulting application
 * OpenAPI spec version: 1.0.0
 */
import type { CreateExtracurricularRequestStatus } from './createExtracurricularRequestStatus';
import type { CreateExtracurricularRequestType } from './createExtracurricularRequestType';

export interface CreateExtracurricularRequest {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  description: string;
  end_date?: string;
  leadership_role?: string;
  /** @minLength 1 */
  name: string;
  organization: string;
  start_date: string;
  status: CreateExtracurricularRequestStatus;
  type: CreateExtracurricularRequestType;
}